Fix FoodPurchase and OtherCashIncome datatypes for months_per_year, t…

+++ apps/baseline/models.py
@@ -2412,13 +2412,15 @@ class FoodPurchase(LivelihoodActivity):
     # This is a float field because data may be captured as "once per week",
     # which equates to "52 per year", which is "4.33 per month".
     times_per_month = models.FloatField(blank=True, null=True, verbose_name=_("Purchases per month"))
-    months_per_year = models.PositiveSmallIntegerField(
+    # This is a float field because BSS data may have fractional months (e.g. 6.4 months).
+    months_per_year = models.FloatField(
         blank=True,
         null=True,
         verbose_name=_("Months per year"),
         help_text=_("Number of months in a year that the product is purchased"),
     )
-    times_per_year = models.PositiveSmallIntegerField(
+    # This is a float field because BSS data may have fractional times (e.g. 103.2 purchases/year).
+    times_per_year = models.FloatField(
         blank=True,
         null=True,
         verbose_name=_("Times per year"),
@@ -2650,7 +2652,8 @@ class OtherCashIncome(LivelihoodActivity):
         verbose_name=_("Payment per time"),
         help_text=_("Amount of money received each time the labor is performed"),
     )
-    people_per_household = models.PositiveSmallIntegerField(
+    # This is a float field because BSS data may have fractional people per household (e.g. 0.5).
+    people_per_household = models.FloatField(
         verbose_name=_("People per household"),
         blank=True,
         null=True,
@@ -2731,13 +2734,15 @@ class OtherPurchase(LivelihoodActivity):
     # This is a float field because data may be captured as "once per week",
     # which equates to "52 per year", which is "4.33 per month".
     times_per_month = models.FloatField(blank=True, null=True, verbose_name=_("Purchases per month"))
-    months_per_year = models.PositiveSmallIntegerField(
+    # This is a float field because BSS data may have fractional months (e.g. 6.4 months).
+    months_per_year = models.FloatField(
         blank=True,
         null=True,
         verbose_name=_("Months per year"),
         help_text=_("Number of months in a year that the product is purchased"),
     )
-    times_per_year = models.PositiveSmallIntegerField(
+    # This is a float field because times_per_month * months_per_year may be fractional (e.g. 51.6 * 4 = 206.4).
+    times_per_year = models.FloatField(
         blank=True,
         null=True,
         verbose_name=_("Times per year"),
